Swift 3.1 and MacOs
I noticed that in the swift language there are practically no lessons or programs for the mackintosh! I'm from Russia. But even on Western resources or in English-language sites or on YouTube. Programming materials for macs are not a class. Only old lessons, when the swift only came out officially. Therefore, I made a simple program for the mac on the last version of this programming language at the moment.